==> questions/01_Prepare_Dataset.ipynb <==
# 01 Pre Augmentation Preparation

- Input: `None`
- Output: `01_prepared_dataset.csv`

Before we can start with the augmentation, we need to prepare the dataset. 

import hashlib
import pandas as pd 

from utils import clean_text

from datasets import load_dataset, concatenate_datasets
from copy import deepcopy
from tqdm import tqdm

iapp    = load_dataset("iapp_wiki_qa_squad")
thaiqa  = load_dataset("thaiqa_squad")
xquad   = load_dataset("xquad", "xquad.th")
tydiqa = load_dataset("khalidalt/tydiqa-goldp", "thai")

# Merge all splits of each dataset
def merge_dataset_splits(dataset):
    splits = list(dataset.keys())
    if len(splits) == 1:
        return dataset[splits[0]]
    else:
        return concatenate_datasets([dataset[split] for split in splits])

iapp = merge_dataset_splits(iapp).to_pandas()
thaiqa = merge_dataset_splits(thaiqa).to_pandas()
xquad = merge_dataset_splits(xquad).to_pandas()
tydiqa = merge_dataset_splits(tydiqa).to_pandas()

## Reformat the Datasets

iapp = iapp[['question', 'context', 'answers']]
thaiqa = thaiqa[['question', 'context', 'answers']]
xquad = xquad[['question', 'context', 'answers']]
tydiqa = tydiqa.rename(columns={'passage_text': 'context', "question_text": "question"})[['question', 'context', 'answers']]

# Datasets require more processing to make them all in the correct format
def get_offset_begin_position(cleaned_context, answers_text, answer_begin_positions):
    new_answer_begin_positions = []

    try:
        for answer_text, answer_begin_position in zip(answers_text, answer_begin_positions):
            # Find all all instances of the answer in cleaned_context
            possible_answer_begin_positions = [i for i in range(len(cleaned_context)) if cleaned_context.startswith(answer_text, i)] 

            if len(possible_answer_begin_positions) == 1:
                new_answer_begin_positions.append(possible_answer_begin_positions[0])
            elif len(possible_answer_begin_positions) > 1:
                # Find the closest answer to the original answer
                closest_answer_begin_position = min(possible_answer_begin_positions, key=lambda x:abs(x-answer_begin_position))
                new_answer_begin_positions.append(closest_answer_begin_position)
            else:
                raise Exception("No answer found in context")
    except:
        print("Error with answer: ", answer_text)
        print("Error with context: ", cleaned_context)
        print("Error with answer begin position: ", answer_begin_position)
        raise
    
    return new_answer_begin_positions

def process_row(row, answer_text_key, answer_start_key):
    # Normalize the text
    new_row = deepcopy(row)
    new_row["context"] = clean_text(row["context"])
    new_row["question"] = clean_text(row["question"], is_question=True)

    new_row["answers"] = {}
    new_row["answers"]["text"] = [clean_text(x) for x in row["answers"][answer_text_key]]

    # Reindex the dataset
    new_row["answers"]["answer_start"] = get_offset_begin_position(new_row["context"], new_row["answers"]["text"], row["answers"][answer_start_key])
    new_row["answers"]["answer_end"] = [x + len(y) for x, y in zip(new_row["answers"]["answer_start"], new_row["answers"]["text"])]

    return new_row

def sanity_check(datasets):
    # Match keys
    for dataset in tqdm(datasets):
        assert list(dataset.columns) == list(datasets[0].columns)
        assert dataset['answers'][0].keys() == datasets[0]['answers'][0].keys()
    
    print("All Keys Matched...")

    # Check theortical answers vs index
    for dataset in datasets:
        for _, row in tqdm(list(dataset.iterrows())):
            for text, begin, end in zip(row["answers"]['text'], row["answers"]['answer_start'], row["answers"]['answer_end']):
                assert text == row['context'][begin:end], f"Theoretical Answer: {text} | Indexed: {row['context'][begin:end]} | Context: {row['context']}"
    
    print("Theortical Answers Matched...")

    # Assert no empty columns
    for dataset in tqdm(datasets):
        assert not dataset.isnull().values.any()
    
    print("No empty columns...")

### Special Scenario for TyDiQA

Some of the following contexts (displayed below and discovered through cleaning & experimentation), have invalid answers—  thus we need to strip them out.
- 480
- 686
- 953
- 2057
- 3177

tydiqa[tydiqa["context"].str.startswith('โจแอนน์ "โจ"')]

tydiqa[tydiqa["context"].str.startswith('โจแอนน์ "โจ"')].iloc[0]["context"]

tydiqa[tydiqa["context"].str.startswith('โจแอนน์ "โจ"')].iloc[1]["question"]

clean_text(tydiqa[tydiqa["context"].str.startswith('โจแอนน์ "โจ"')].iloc[0]["context"])[150:] # This answer does not make sense at all

tydiqa[tydiqa["context"].str.startswith("พระสิทธัตถพุทธเจ้า")] # Answer is incorrectly formatted

#### Find all invalid answers

tydiqa[tydiqa["answers"].apply(lambda x: any(["}'>" in y for y in x["text"]]))]

tydiqa[tydiqa["answers"].apply(lambda x: any([":[" in y for y in x["text"]]))]

tydiqa[tydiqa["answers"].apply(lambda x: any(["=" in y for y in x["text"]]))]

tydiqa[tydiqa["answers"].apply(lambda x: any(["[" in y for y in x["text"]]))]

tydiqa = tydiqa.drop([480, 686, 953, 2057, 3177])
tydiqa = tydiqa[tydiqa["answers"].apply(lambda x: not any(["}'>" in y for y in x["text"]]))]
tydiqa = tydiqa[tydiqa["answers"].apply(lambda x: not any([":[" in y for y in x["text"]]))]
tydiqa = tydiqa[tydiqa["answers"].apply(lambda x: not any(["=" in y for y in x["text"]]))]
tydiqa = tydiqa[tydiqa["answers"].apply(lambda x: not any(["[" in y for y in x["text"]]))]

### Apply cleaning process to dataset

iapp = iapp.apply(lambda x: process_row(x, "text", "answer_start"), axis=1)
thaiqa = thaiqa.apply(lambda x: process_row(x, "answer", "answer_begin_position"), axis=1)
xquad = xquad.apply(lambda x: process_row(x, "text", "answer_start"), axis=1)
tydiqa = tydiqa.apply(lambda x: process_row(x, "text", "start_byte"), axis=1)

sanity_check([iapp, thaiqa, xquad, tydiqa])

## Merge the Datasets

iapp["source"] = "iapp"
thaiqa["source"] = "thaiqa"
xquad["source"] = "xquad"
tydiqa["source"] = "tydiqa"

final = pd.concat([iapp, thaiqa, xquad, tydiqa], ignore_index=True)
final

# Drop duplicates to avoid accidentally leaking data into the test set
print(f"Before dropping duplicates: {len(final)}")
print(f"Values per source: {final['source'].value_counts()}")
print("")

final = final.drop_duplicates(subset=["context", "question"], keep="first").reset_index(drop=True)

print(f"After dropping duplicates: {len(final)}")
print(f"Values per source: {final['source'].value_counts()}")

# Generate unique ids for each row by hashing the context and question
final["id"] = final.apply(lambda x: hashlib.sha256((x["context"] + x["question"]).encode("utf-8")).hexdigest(), axis=1)

final.to_csv("data/01_prepare_dataset.csv", index=False)
