==> questions/04_Paraphrase_QCPG.ipynb <==
# 04 QCPG Paraphrase
- Input: `03_augment_llm.csv`
- Output: `04_qcpg_paraphrase.csv`

import gc
import platform
import pandas as pd

import torch

from datasets import Dataset
from transformers import pipeline
from transformers.pipelines.pt_utils import KeyDataset

from tqdm.notebook import tqdm

# If on mac default to mps, otherwise cuda, then cpu
if torch.cuda.is_available():
    device = torch.device("cuda")
else:
    # Check if on mac
    if platform.system() == "Darwin":
        device = torch.device("mps")
    else:
        device = torch.device("cpu")
device

## Load data

df = pd.read_csv("data/03_augment_llm.csv")
df

pipe = pipeline('text2text-generation', model=f'ibm/qcpg-questions', device=device)

def run_augmentation(df, column_name, sem, lex, syn):
    df = df.copy(deep=True)
    df[column_name] = f"COND_SEMANTIC_SIM_{int(sem*100)} COND_LEXICAL_DIV_{int(lex*100)} COND_SYNTACTIC_DIV_{int(syn*100)} " + df[column_name]
    qcpg_augmented = []

    dataset = Dataset.from_pandas(df)

    for out in tqdm(pipe(KeyDataset(dataset, column_name), batch_size=64), desc=f"Running QCPG", total=len(df)):
        qcpg_augmented.append(out[0]["generated_text"])
    
    gc.collect()
    
    return qcpg_augmented

settings_values = [0.2, 0.5, 0.8] 
columns = ["en_aug", "en_llm_gec_aug"]
for column in tqdm(columns, desc="Augmenting Columns"):
    for value in tqdm(settings_values, desc="Augmenting Column Values"):
        if column == "en_aug":
            target_name = f"en_qcpg_{value}_aug"
        else:
            target_name = f"en_qcpg_{value}_{'_'.join(column.split('_')[1:-1])}_aug"
        
        df[target_name] = run_augmentation(df, column, value, value, value)

# Sanity check for no empty rows
assert df.isnull().sum().sum() == 0
df

# Export and save QCPG
df.to_csv("data/04_paraphrase_qcpg.csv", index=False)
