==> questions/07_Benchmark_ReQA.ipynb <==
# Benchmark ReQA
- Input: `06_calculate_distance.parquet`

import pandas as pd
import numpy as np
from tqdm.notebook import tqdm
from utils import encode_in_batch

import tensorflow_text
import tensorflow as tf
import tensorflow_hub as hub

data = pd.read_parquet('data/06_calculate_distance.parquet')
data

# Add None so that the first benchmark only consists of the original questions
all_augment_cols = [None] + [col for col in data.columns if col.startswith('th_')]
all_augment_cols

model_embed = hub.load("https://tfhub.dev/google/universal-sentence-encoder-multilingual/3")

def evaluate(
    question_id, question_all, context_id, context_all, mrr_rank=10, status=True
):
    top_1 = 0
    top_5 = 0
    top_10 = 0
    mrr_score = 0
    context_id = np.array(context_id)
    sim_score = np.inner(question_all, context_all)
    if status == True:
        status_bar = enumerate(tqdm(sim_score))
    else:
        status_bar = enumerate(sim_score)
    for idx, sim in status_bar:
        index = np.argsort(sim)[::-1]
        index_edit = [context_id[x] for x in index]
        try:
            idx_search = list(index_edit).index(question_id[idx])
        except:
            idx_search = 999999
        if idx_search == 0:
            top_1 += 1
            top_5 += 1
            top_10 += 1
        elif idx_search < 5:
            top_5 += 1
            top_10 += 1
        elif idx_search < 10:
            top_10 += 1
        if idx_search < mrr_rank:
            mrr_score += 1 / (idx_search + 1)
    return (
        top_1 / len(question_all),
        top_5 / len(question_all),
        top_10 / len(question_all),
        mrr_score / len(question_all),
    )

def benchmark_single(data, question_col="question", aug_col=None, model_embed=model_embed, filter_zero_dist=True):
    # Check if column has a distance counterpart
    questions = data[question_col].values

    question_all = encode_in_batch(model_embed, questions)
    context_all = encode_in_batch(model_embed, data["context"].values)

    question_ids = data["id"].values
    context_ids = data["id"].values

    if aug_col:
        augments = data[aug_col].values

        if filter_zero_dist:
            distances = data[aug_col.replace("th_", "dis_")].values
            augments = augments[distances > 0]
            augment_ids = data["id"].values[distances > 0]
        else:
            augment_ids = data["id"].values
        
        augment_all = encode_in_batch(model_embed, augments)
        
        # Augment into context
        context_all = np.concatenate([context_all, augment_all], axis=0)
        context_ids = np.concatenate([context_ids, augment_ids], axis=0)

    top_1, top_5, top_10, mrr_score = evaluate(
        question_ids, question_all, context_ids, context_all, status=True
    )
    return top_1, top_5, top_10, mrr_score

results = []

for col in tqdm(all_augment_cols):
    top_1, top_5, top_10, mrr_score = benchmark_single(data, aug_col=col)
    print(f"{col}: top_1: {top_1}, top_5: {top_5}, top_10: {top_10}, mrr_score: {mrr_score}")
    results.append({"name": col, "top_1": top_1, "top_5": top_5, "top_10": top_10, "mrr_score": mrr_score})

results = pd.DataFrame(results)
results

results.to_csv("results/07_benchmark_reqa.csv", index=False)
